Modify: move Read more button into a separate component.
